pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "changescout"
version = "0.1.0"
description = "Deterministic monitoring pipeline for canton scoped infrastructure change lead generation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
]

[project.scripts]
changescout = "changescout.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"